+++ src/module/vehicles/vehicle.services.ts
@@ -53,6 +53,12 @@ const updateVehicle = async(id:number, data:updatedInfo)=>{
 }
 
 const deleteVehicle = async(id:number)=>{
+    const booking = await pool.query(`SELECT * FROM bookings WHERE vehicle_id=$1`,[id]);
+
+    if(booking.rows.length !== 0){
+        return false;
+    }
+
     const result = await pool.query(`DELETE FROM vehicles WHERE id=$1`,[id]);
     return result;
 }
